PWN-7591 - comment fixes
