+++ app/src/main/java/org/p2p/wallet/jupiter/ui/main/SwapTokenRateLoader.kt
@@ -9,27 +9,6 @@ import kotlinx.coroutines.flow.flow
 import org.p2p.wallet.jupiter.interactor.model.SwapTokenModel
 import org.p2p.wallet.jupiter.repository.tokens.JupiterSwapTokensRepository
 
-class SwapTokenRateNotFound(
-    token: SwapTokenModel
-) : Throwable() {
-    override val message: String = buildString {
-        append("Token rate not found for token ")
-        append(token.tokenName)
-        append("; ")
-        append(token.mintAddress)
-        append("; ")
-        append(
-            when (token) {
-                is SwapTokenModel.JupiterToken -> token.coingeckoId
-                is SwapTokenModel.UserToken -> null
-            }
-        )
-        append("; ")
-        append(token::class.simpleName)
-        append("; ")
-    }
-}
-
 class SwapTokenRateLoader(
     private val swapTokensRepository: JupiterSwapTokensRepository,
 ) {
